Initial work to properly use the MVVM pattern
Created the services and scaffolding for the basic architecture.
TODO: Create ViewModels for every view to separate concerns.
Moved core models to separate project.
In the process of switching to SyncFusion Data Grid for folder list
(currently displaying both versions in the main view so I have a
reference for styling).
Moved to .net10